# capcloud-api/public/exports/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "北京市建筑垃圾公开查询接口（非官方整理）",
    "version": "2026-09-05",
    "description": "根据原站前端整理并匿名抽样核验。旧平台存在关停风险；本文件不是官方接口或服务承诺。"
  },
  "servers": [
    {
      "url": "http://ztxn.capcloud.com.cn:8080/dregs_service-dev"
    }
  ],
  "paths": {
...
    "/dictionary/area/dictionaryValue": {
      "get": {
        "operationId": "districts",
        "summary": "审批区字典",
        "security": [],
        "parameters": [
          {
            "name": "area",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "conferenceType"
              ],
              "default": "conferenceType"
            },
            "description": "对应截图入口的固定查询条件"
          }
        ],
        "responses": {
          "200": {
            "description": "须同时检查 success、code；HTTP 200 不代表业务成功。分页列表位于 result.rows 和 result.total。",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "result": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "dicname": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      }
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    }
  }
}